[WJ] Extract utility functions
